+++ packages/@headlessui-react/src/components/listbox/listbox.tsx
@@ -756,7 +756,7 @@ let Options = forwardRefWithAs(function Options<
   let usesOpenClosedState = useOpenClosed()
   let visible = (() => {
     if (usesOpenClosedState !== null) {
-      return usesOpenClosedState === State.Open
+      return (usesOpenClosedState & State.Open) === State.Open
     }
 
     return data.listboxState === ListboxStates.Open